refactor: Lotto class를 Model package로 이동
