+++ src/main/java/Model/Lotto.java
@@ -1,4 +1,4 @@
-package lotto;
+package Model;
 
 import java.util.List;
